Refine test cases for MathUtility
