+++ app/src/test/java/mpc/project/util/KeyTest.java
@@ -1,12 +1,12 @@
 package mpc.project.util;
 
-import junit.framework.TestCase;
 import org.junit.Test;
 
 import java.io.IOException;
 import java.math.BigInteger;
 
-public class KeyTest extends TestCase {
+public class KeyTest {
+    @Test
     public void testToPKCS1PublicString() {
         Key key = new Key();
         key.setN(BigInteger.valueOf(57761551163L));
